bitcoin: move taproot test to tests/
Move bip_341_tests to tests/ because it include_str's
data from tests/, which we exclude when publishing.
